openlabcmd/requirements.txt
enum;python_version=='2.7'
kazoo
pbr>=1.3
prettytable
PyYAML>=5.1
